chore: ultra-clean artifacts
